+++ Fusepp-2.0/ltrmount/LTRRepositoryWrapper.h
@@ -17,7 +17,7 @@ public:
 
 	static std::vector<std::string> readdir(const char *path);
 
-	static int read(const char *path, char *buf, size_t size, size_t offset);
+	static size_t read(const char *path, char *buf, size_t size, size_t offset);
 };
